requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4
pdfplumber==0.11.4
Pillow==11.1.0
groq==0.13.0
